+++ pythran/pythonic/numpy/ndarray/flatten.hpp
@@ -10,7 +10,7 @@ namespace pythonic { namespace numpy  {
         template<class T, size_t N>
             types::ndarray<T, 1>
             flatten(types::ndarray<T,N> const& a) {
-                return types::ndarray<T, 1>(a.mem, types::array<long, 1>{{a.flat_size()}});
+                return types::ndarray<T, 1>(a.mem, types::array<long, 1>{{static_cast<long>(a.flat_size())}});
             }
 
         NUMPY_EXPR_TO_NDARRAY0(flatten);
